Switch scanner pipeline to 8-bit and forward scan frames to CameraView
- Frame::data changed from uint16_t to uint8_t throughout (ICamera.h,
  LaserLineExtractor.h/.cpp); qImageToScannerFrame drops the <<8 shift
- ExtractorParams::threshold narrowed to uint8_t, default 50 (was 500 in
  16-bit space); spinbox range changed to 0–255
- Histogram X-axis updated to 8-bit (0–255, ticks at multiples of 32);
  Y-axis switched from log to linear scale; left margin tightened
- ScannerTab emits previewFrameReady(QImage) after each scan-step grab;
  MainWindow wires it to CameraView so the laser line is visible during
  scanning (acquisition thread is stopped during scan)

Scanner calibration still untested.

+++ src/scanner/interfaces/ICamera.h
@@ -21,16 +21,15 @@ struct CameraParams {
     AcquisitionMode mode{AcquisitionMode::SoftwareTrigger};
 };
 
-// Frame data is always 16-bit monochrome, row-major.
-// IDSCamera expands Mono8→16 (×256) and Mono12→16 (<<4) before returning.
+// Frame data is 8-bit monochrome, row-major.
 struct Frame {
-    std::vector<uint16_t> data;
+    std::vector<uint8_t> data;
     int width{0};
     int height{0};
 
     bool valid() const noexcept { return !data.empty() && width > 0 && height > 0; }
 
-    uint16_t pixel(int col, int row) const noexcept {
+    uint8_t pixel(int col, int row) const noexcept {
         return data[static_cast<std::size_t>(row * width + col)];
     }
 };
